fix(evaluation): extract group, fix phases crash
Split the inline Model/Prompt/Skill A/B rendering block out of
Evaluation.svelte into a reusable ExperimentGroup component, sharing
formatting/classification helpers via lib/evaluation-format.ts. Add a
stable experiment-caveat--dominant class so low-sample badges and
guardrail breaches read louder than the metric they qualify.

Also fix a null-deref: PhaseReport.Phases marshals to JSON null (not
[]) when a cohort has no landed-task phase data yet, e.g. a fresh
install. Evaluation.svelte indexed phases.phases directly and crashed
before ever reaching the new Model/Prompt/Skill empty states.

The backend Kind discriminator and byExperimentKind grouping this
task's plan called for already landed via #1361; this change only
covers the remaining frontend piece.
